third_party/blink/web_tests/http/tests/security/powerfulFeatureRestrictions/device-orientation-handler-not-fired-on-insecure-origin.html

<!DOCTYPE html>
<title>Device Sensor Events on Insecure Origin. Codifies expected behavior once https://github.com/w3c/deviceorientation/issues/47 is resolved.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script src="/resources/sensor-helpers.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/services/device/public/mojom/sensor_provider.mojom.js"></script>
<script>

// Cannot use `step_timeout()` because we need the `sensor_test` infrastructure below,
// which, however, is using `promise_test` internally. Cannot use `promise_rejects`
// either, as `sensor_test` does not expose the `test_object`.
function waitForLackOfEvent(eventName) {
  return new Promise((resolve, reject) => {
    window.addEventListener(eventName, reject);
    window.setTimeout(() => {
      window.removeEventListener(eventName, reject);
      resolve();
    }, 1000);
  });
}

if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
  window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
} else {

  test(() => {
    assert_false('DeviceMotionEvent' in window);
    assert_false('DeviceOrientationEvent' in window);
    assert_false('DeviceMotionEventAcceleration' in window);
    assert_false('DeviceMotionEventRotationRate' in window);
  }, 'Event interfaces are not exposed on `window`.');

  sensor_test(sensorProvider => {
    const FAKE_ACCELERATION_DATA = [1, 2, 3];
    const FAKE_LINEAR_ACCELERATION_DATA = [4, 5, 6];
    const FAKE_GYROSCOPE_DATA = [7, 8, 9];
    setMockSensorDataForType(sensorProvider, device.mojom.SensorType.ACCELEROMETER, FAKE_ACCELERATION_DATA);
    setMockSensorDataForType(sensorProvider, device.mojom.SensorType.LINEAR_ACCELERATION, FAKE_LINEAR_ACCELERATION_DATA);
    setMockSensorDataForType(sensorProvider, device.mojom.SensorType.GYROSCOPE, FAKE_GYROSCOPE_DATA);

    return waitForLackOfEvent('devicemotion');
  }, 'DeviceMotionEvent never fires.');

  sensor_test(sensorProvider => {
    const FAKE_ORIENTATION_DATA = [1.1, 2.2, 3.3];
    setMockSensorDataForType(sensorProvider, device.mojom.SensorType.RELATIVE_ORIENTATION_EULER_ANGLES, FAKE_ORIENTATION_DATA);

    return waitForLackOfEvent('deviceorientation');
  }, 'DeviceOrientationEvent never fires.');
}

</script>
